+++ h2integrate/core/supported_models.py
@@ -42,6 +42,10 @@ from h2integrate.converters.nitrogen.simple_ASU import SimpleASUCostModel, Simpl
 from h2integrate.converters.wind.wind_plant_ard import ArdWindPlantModel
 from h2integrate.resource.solar.openmeteo_solar import OpenMeteoHistoricalSolarResource
 from h2integrate.storage.simple_generic_storage import SimpleGenericStorage
+from h2integrate.converters.hydrogen.h2_fuel_cell import (
+    H2FuelCellCostModel,
+    H2FuelCellPerformanceModel,
+)
 from h2integrate.converters.hydrogen.wombat_model import WOMBATElectrolyzerModel
 from h2integrate.converters.steel.steel_eaf_plant import (
     HydrogenEAFPlantCostComponent,
@@ -182,6 +186,8 @@ supported_models = {
     "BasicElectrolyzerCostModel": BasicElectrolyzerCostModel,
     "CustomElectrolyzerCostModel": CustomElectrolyzerCostModel,
     "WOMBATElectrolyzerModel": WOMBATElectrolyzerModel,
+    "H2FuelCellPerformanceModel": H2FuelCellPerformanceModel,
+    "H2FuelCellCostModel": H2FuelCellCostModel,
     "SimpleASUCostModel": SimpleASUCostModel,
     "SimpleASUPerformanceModel": SimpleASUPerformanceModel,
     "HOPPComponent": HOPPComponent,